requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langwatch>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
